votable: avoid implicit narrowing casts
Be more careful with longs and ints, avoiding implicit narrowing casts
(compound assignment expression semantics).

These changes prevent [lossy-conversions] warnings from Java 21.

+++ votable/src/main/uk/ac/starlink/votable/Decoder.java
@@ -235,7 +235,7 @@ abstract class Decoder {
             isScalar = true;
         }
         else if ( arraysize[ ndim - 1 ] > 0 ) {
-            int nel = 1;
+            long nel = 1;
             for ( int i = 0; i < ndim; i++ ) {
                 nel *= arraysize[ i ];
             }
